net/http/internal/httpcommon: deduplicate gzip reader implementation
net/http and net/http/internal/http2 each carry a copy of a pooled gzip
reader implementation. Unify it by placing it in
net/http/internal/httpcommon.

The new implementation in internal/httpcommon is largely the same,
except for dropping protocol-specific prefix (e.g., "http2:") in its
error, standardizing on using fs.ErrClosed (net/http uses an unexported
errReadOnClosedResBody) when Read is called after Close, and accepting
io.ReadCloser (net/http uses bodyEOFSignal) as the body.

This is moved to net/http/internal/httpcommon, which is kept in sync
with x/net, so HTTP/3 in x/net can use it too.

For #81473

Reviewed-on: https://go-review.googlesource.com/c/go/+/831344
LUCI-TryBot-Result: golang-scoped@luci-project-accounts.iam.gserviceaccount.com <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
