+++ src/net/http/transport.go
@@ -11,8 +11,6 @@ package http
 
 import (
 	"bufio"
-	"compress/flate"
-	"compress/gzip"
 	"container/list"
 	"context"
 	"crypto/tls"
@@ -26,6 +24,7 @@ import (
 	"net/http/httptrace"
 	"net/http/internal"
 	"net/http/internal/ascii"
+	"net/http/internal/httpcommon"
 	"net/textproto"
 	"net/url"
 	"reflect"
@@ -2567,7 +2566,7 @@ func (pc *persistConn) readLoop() {
 
 		resp.Body = body
 		if rc.addedGzip && ascii.EqualFold(resp.Header.Get("Content-Encoding"), "gzip") {
-			resp.Body = &gzipReader{body: body}
+			resp.Body = &httpcommon.GzipReader{Body: body}
 			resp.Header.Del("Content-Encoding")
 			resp.Header.Del("Content-Length")
 			resp.ContentLength = -1
@@ -3212,7 +3211,6 @@ type bodyEOFSignal struct {
 }
 
 var errReadOnClosedResBody = errors.New("http: read on closed response body")
-var errConcurrentReadOnResBody = errors.New("http: concurrent read on response body")
 
 func (es *bodyEOFSignal) Read(p []byte) (n int, err error) {
 	es.mu.Lock()
@@ -3263,117 +3261,6 @@ func (es *bodyEOFSignal) condfn(err error) error {
 	es.fn = nil
 	es.earlyCloseFn = nil
 	return fn(err)
-}
-
-// gzipReader wraps a response body so it can lazily
-// get gzip.Reader from the pool on the first call to Read.
-// After Close is called it puts gzip.Reader to the pool immediately
-// if there is no Read in progress or later when Read completes.
-type gzipReader struct {
-	_    incomparable
-	body *bodyEOFSignal // underlying HTTP/1 response body framing
-	mu   sync.Mutex     // guards zr and zerr
-	zr   *gzip.Reader   // stores gzip reader from the pool between reads
-	zerr error          // sticky gzip reader init error or sentinel value to detect concurrent read and read after close
-}
-
-type eofReader struct{}
-
-func (eofReader) Read([]byte) (int, error) { return 0, io.EOF }
-func (eofReader) ReadByte() (byte, error)  { return 0, io.EOF }
-
-var gzipPool = sync.Pool{New: func() any { return new(gzip.Reader) }}
-
-// gzipPoolGet gets a gzip.Reader from the pool and resets it to read from r.
-func gzipPoolGet(r io.Reader) (*gzip.Reader, error) {
-	zr := gzipPool.Get().(*gzip.Reader)
-	if err := zr.Reset(r); err != nil {
-		gzipPoolPut(zr)
-		return nil, err
-	}
-	return zr, nil
-}
-
-// gzipPoolPut puts a gzip.Reader back into the pool.
-func gzipPoolPut(zr *gzip.Reader) {
-	// Reset will allocate bufio.Reader if we pass it anything
-	// other than a flate.Reader, so ensure that it's getting one.
-	var r flate.Reader = eofReader{}
-	zr.Reset(r)
-	gzipPool.Put(zr)
-}
-
-// acquire returns a gzip.Reader for reading response body.
-// The reader must be released after use.
-func (gz *gzipReader) acquire() (*gzip.Reader, error) {
-	gz.mu.Lock()
-	defer gz.mu.Unlock()
-	if gz.zerr != nil {
-		return nil, gz.zerr
-	}
-	if gz.zr == nil {
-		// gzipPoolGet might block indefinitely since it reads the gzip header.
-		// Therefore, drop mu temporarily when using gzipPoolGet.
-		// We set zerr to errConcurrentReadOnResBody to prevent concurrent read
-		// even when mu is temporarily dropped.
-		gz.zerr = errConcurrentReadOnResBody
-		gz.mu.Unlock()
-		zr, err := gzipPoolGet(gz.body)
-		gz.mu.Lock()
-		// Guard against Close being called while gzipPoolGet is running.
-		if gz.zerr != errConcurrentReadOnResBody {
-			if zr != nil {
-				gzipPoolPut(zr)
-			}
-			return nil, gz.zerr
-		}
-		gz.zr, gz.zerr = zr, err
-		if gz.zerr != nil {
-			return nil, gz.zerr
-		}
-	}
-	ret := gz.zr
-	gz.zr, gz.zerr = nil, errConcurrentReadOnResBody
-	return ret, nil
-}
-
-// release returns the gzip.Reader to the pool if Close was called during Read.
-func (gz *gzipReader) release(zr *gzip.Reader) {
-	gz.mu.Lock()
-	defer gz.mu.Unlock()
-	if gz.zerr == errConcurrentReadOnResBody {
-		gz.zr, gz.zerr = zr, nil
-	} else { // errReadOnClosedResBody
-		gzipPoolPut(zr)
-	}
-}
-
-// close returns the gzip.Reader to the pool immediately or
-// signals release to do so after Read completes.
-func (gz *gzipReader) close() {
-	gz.mu.Lock()
-	defer gz.mu.Unlock()
-	if gz.zerr == nil && gz.zr != nil {
-		gzipPoolPut(gz.zr)
-		gz.zr = nil
-	}
-	gz.zerr = errReadOnClosedResBody
-}
-
-func (gz *gzipReader) Read(p []byte) (n int, err error) {
-	zr, err := gz.acquire()
-	if err != nil {
-		return 0, err
-	}
-	defer gz.release(zr)
-
-	return zr.Read(p)
-}
-
-func (gz *gzipReader) Close() error {
-	gz.close()
-
-	return gz.body.Close()
 }
 
 type tlsHandshakeTimeoutError struct{}
